breaking: moving token refresh logic into SupabaseAuth
